+++ aasrp/views/datatables.py
@@ -12,11 +12,10 @@ from allianceauth.framework.datatables import DataTablesView
 from allianceauth.services.hooks import get_extension_logger
 
 # AA SRP
-from aasrp import __title__
 from aasrp.models import SrpRequest
-from aasrp.providers import AppLogger
+from aasrp.providers.applogger import AppLogger
 
-logger = AppLogger(my_logger=get_extension_logger(name=__name__), prefix=__title__)
+logger = AppLogger(my_logger=get_extension_logger(name=__name__))
 
 
 class OwnSrpRequestsView(PermissionRequiredMixin, DataTablesView):